Support directory index files
If a component is not found for a path, assume path is a directory and
search for directory index files.

For now we hardcode the files to 'index.html' and 'index.htm' but it could
be a configuration.

Also the search order is wrong with multiple components. A proper solution
would be to move this to Interp.pm. Patches welcome :).

Signed-off-by: Pedro Melo <melo@simplicidade.org>